pyproject.toml
[project]
name = "dataAnalysis3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.2.2",
    "httpx>=0.28.1",
    "langchain>=0.3.25",
    "langchain-mcp-adapters>=0.0.11",
    "langchain-openai>=0.3.16",
    "langgraph>=0.4.3",
    "mcp-server-git>=2025.1.14",
    "mcp[cli]>=1.8.0",
    "pandas>=2.2.3",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "uv>=0.7.3",
    "nest-asyncio>=1.5.6",
    "ipython>=8.0.0",
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=2.0.0",
    "plotly>=5.0.0",
    "tabulate>=0.9.0",
    "grandalf>=0.8",
    "langchain-anthropic>=0.3.19",
    "langchain-groq>=0.3.7",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
